Added tests. There are failures
