+++ src/test/kotlin/eu/europa/ec/eudi/sdjwt/dsl/def/DefinitionBasedSdJwtVcValidatorTest.kt
@@ -26,6 +26,7 @@ import eu.europa.ec.eudi.sdjwt.vc.DefinitionBasedSdJwtVcValidator
 import eu.europa.ec.eudi.sdjwt.vc.DefinitionBasedValidationResult
 import eu.europa.ec.eudi.sdjwt.vc.DefinitionViolation
 import kotlin.test.Test
+import kotlin.test.assertContentEquals
 import kotlin.test.assertEquals
 import kotlin.test.assertIs
 
@@ -197,6 +198,48 @@ class DefinitionBasedSdJwtVcValidatorTest {
                 }
             },
     )
+
+    @Test
+    fun arrayTestHappyPath() = AddressDefinition.shouldConsiderValid(
+        sdJwtObject = sdJwt {
+            sdArrClaim("addresses") {
+                sdObjClaim {
+                    sdClaim("country", "country0")
+                }
+                sdObjClaim {
+                    sdClaim("country", "country1")
+                }
+                sdObjClaim {
+                    sdClaim("country", "country2")
+                }
+            }
+        },
+    )
+
+    @Test
+    fun detectIncorrectlyDisclosedClaimInsideArray() {
+        val sdJwt = sdJwt {
+            sdArrClaim("addresses") {
+                objClaim {
+                    sdClaim("country", "country0")
+                }
+                sdObjClaim {
+                    sdClaim("country", "country1")
+                }
+                objClaim {
+                    sdClaim("country", "country2")
+                }
+            }
+        }
+
+        val errors = AddressDefinition.shouldConsiderInvalid(sdJwtObject = sdJwt)
+        val expectedErrors =
+            listOf(
+                ClaimPath.claim("addresses").arrayElement(0),
+                ClaimPath.claim("addresses").arrayElement(2),
+            ).map { DefinitionViolation.IncorrectlyDisclosedClaim(it) }
+        assertContentEquals(expectedErrors, errors)
+    }
 
     private val default: DefinitionBasedSdJwtVcValidator = DefinitionBasedSdJwtVcValidator.UsingFold
