+++ src/test/kotlin/eu/europa/ec/eudi/sdjwt/dsl/def/SdJwtObjectDefinitionFromSdJwtVcMetadataTest.kt
@@ -22,8 +22,6 @@ import eu.europa.ec.eudi.sdjwt.dsl.sdjwt.def.SdJwtDefinition
 import eu.europa.ec.eudi.sdjwt.dsl.sdjwt.def.claimPaths
 import eu.europa.ec.eudi.sdjwt.dsl.sdjwt.def.fromSdJwtVcMetadata
 import eu.europa.ec.eudi.sdjwt.vc.ResolvedTypeMetadata
-import eu.europa.ec.eudi.sdjwt.vc.SdJwtVcTypeMetadata
-import kotlinx.serialization.json.Json
 import kotlin.test.Test
 import kotlin.test.assertContentEquals
 import kotlin.test.assertEquals
@@ -31,28 +29,33 @@ import kotlin.test.assertIs
 
 class SdJwtObjectDefinitionFromSdJwtVcMetadataTest {
 
+    private fun commonChecks(
+        resolvedTypeMetadata: ResolvedTypeMetadata,
+        sdJwtDefinition: SdJwtDefinition,
+    ) {
+        val vctMetadata = sdJwtDefinition.metadata
+        assertEquals(resolvedTypeMetadata.vct, vctMetadata.vct)
+        assertEquals(resolvedTypeMetadata.name, vctMetadata.name)
+        assertEquals(resolvedTypeMetadata.description, vctMetadata.description)
+        assertEquals(resolvedTypeMetadata.display, vctMetadata.display)
+
+        val expectedClaimPaths =
+            resolvedTypeMetadata.claims.map { it.path }
+        val claimPaths = sdJwtDefinition.claimPaths()
+        assertContentEquals(expectedClaimPaths, claimPaths)
+    }
+
     @Test
-    fun test() {
-        val pidSdJwtVcTypeMetadata = sdJwtVcTypeMetadata(pidMeta).resolve()
-        val pidDefinition = SdJwtDefinition.fromSdJwtVcMetadata(pidSdJwtVcTypeMetadata)
+    fun checkPidDefinition() {
+        val sdJwtVcTypeMetadata = sdJwtVcTypeMetadata(pidMeta).resolve()
+        val sdJwtVcDefinition = SdJwtDefinition.fromSdJwtVcMetadata(sdJwtVcTypeMetadata)
 
-        val vctMetadata = pidDefinition.metadata
-        assertEquals(pidSdJwtVcTypeMetadata.vct, vctMetadata.vct)
-        assertEquals(pidSdJwtVcTypeMetadata.name, vctMetadata.name)
-        assertEquals(pidSdJwtVcTypeMetadata.description, vctMetadata.description)
-        assertEquals(pidSdJwtVcTypeMetadata.display, vctMetadata.display)
-
-        assertContentEquals(
-            pidSdJwtVcTypeMetadata.claims.map { it.path },
-            pidDefinition.claimPaths(),
-        )
-
-        val nationalities = pidDefinition.content["nationalities"]
+        val nationalities = sdJwtVcDefinition.content["nationalities"]
         assertIs<Disclosable.AlwaysSelectively<DisclosableValue.Arr<String, AttributeMetadata>>>(nationalities)
         assertEquals(1, nationalities.value.value.content.size)
         assertIs<Disclosable.AlwaysSelectively<DisclosableValue.Id<String, AttributeMetadata>>>(nationalities.value.value.content.first())
 
-        val address = pidDefinition.content["address"]
+        val address = sdJwtVcDefinition.content["address"]
         assertIs<Disclosable.AlwaysSelectively<DisclosableValue.Obj<String, AttributeMetadata>>>(address)
         listOf(
             "street_address",
@@ -67,7 +70,7 @@ class SdJwtObjectDefinitionFromSdJwtVcMetadataTest {
             assertIs<Disclosable.AlwaysSelectively<DisclosableValue.Id<String, AttributeMetadata>>>(attribute)
         }
 
-        val placeOfBirth = pidDefinition.content["place_of_birth"]
+        val placeOfBirth = sdJwtVcDefinition.content["place_of_birth"]
         assertIs<Disclosable.AlwaysSelectively<DisclosableValue.Obj<String, AttributeMetadata>>>(placeOfBirth)
         listOf(
             "locality",
@@ -77,467 +80,23 @@ class SdJwtObjectDefinitionFromSdJwtVcMetadataTest {
             val attribute = placeOfBirth.value.value.content[path]
             assertIs<Disclosable.AlwaysSelectively<DisclosableValue.Id<String, AttributeMetadata>>>(attribute)
         }
+
+        commonChecks(sdJwtVcTypeMetadata, sdJwtVcDefinition)
+    }
+
+    @Test
+    fun checkAddressDefinition() {
+        val sdJwtVcTypeMetadata = sdJwtVcTypeMetadata(addressMeta).resolve()
+        val sdJwtVcDefinition = SdJwtDefinition.fromSdJwtVcMetadata(sdJwtVcTypeMetadata)
+
+        val addressArrayDef = sdJwtVcDefinition.content["addresses"]
+        assertIs<Disclosable.AlwaysSelectively<DisclosableValue.Arr<String, AttributeMetadata>>>(addressArrayDef)
+        assertEquals(1, addressArrayDef.value.value.content.size)
+        val addressDef =
+            assertIs<Disclosable.AlwaysSelectively<DisclosableValue.Obj<String, AttributeMetadata>>>(
+                addressArrayDef.value.value.content.first(),
+            )
+
+        commonChecks(sdJwtVcTypeMetadata, sdJwtVcDefinition)
     }
 }
-
-internal val PidDefinition: SdJwtDefinition by lazy {
-    val sdJwtVcTypeMetadata = sdJwtVcTypeMetadata(pidMeta)
-    val resolvedTypeMetadata = sdJwtVcTypeMetadata.resolve()
-    SdJwtDefinition.fromSdJwtVcMetadata(resolvedTypeMetadata)
-}
-private fun SdJwtVcTypeMetadata.resolve(): ResolvedTypeMetadata {
-    return ResolvedTypeMetadata(
-        vct = vct,
-        name = name,
-        description = description,
-        display = checkNotNull(display).value,
-        claims = checkNotNull(claims),
-        schemas = schema?.let { listOf(it) }.orEmpty(),
-    )
-}
-
-@Suppress("SameParameterValue")
-private fun sdJwtVcTypeMetadata(json: String): SdJwtVcTypeMetadata = Json.decodeFromString(json)
-
-private val pidMeta = """
-    {
-      "vct": "urn:eudi:pid:1",
-      "name": "Type Metadata for Person Identification Data",
-      "display": [
-        {
-          "lang": "en",
-          "name": "PID",
-          "description": "Person Identification Data"
-        }
-      ],
-      "claims": [
-        {
-          "path": [
-            "family_name"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Family Name(s)"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "given_name"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Given Name(s)"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "birthdate"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Birth Date"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "place_of_birth"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Birth Place"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "place_of_birth",
-            "locality"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Locality"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "place_of_birth",
-            "region"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Region"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "place_of_birth",
-            "country"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Country"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "nationalities"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Nationality"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "nationalities",
-            null
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "address"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Address"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "address",
-            "house_number"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "House Number"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "address",
-            "street_address"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Street"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "address",
-            "locality"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Locality"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "address",
-            "region"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Region"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "address",
-            "postal_code"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Postal Code"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "address",
-            "country"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Country"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "address",
-            "formatted"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Full Address"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "personal_administrative_number"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Personal Administrative Number"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "picture"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Portrait Image"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "birth_family_name"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Birth Family Name(s)"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "birth_given_name"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Birth Given Name(s)"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "sex"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Sex"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "email"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Email Address"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "phone_number"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Mobile Phone Number"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "date_of_expiry"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Expiry Date"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "issuing_authority"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Issuing Authority"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "issuing_country"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Issuing Country"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "document_number"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Document Number"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "issuing_jurisdiction"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Issuing Jurisdiction"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "date_of_issuance"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Issuance Date"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "age_equal_or_over"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Age Equal or Over"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "age_equal_or_over",
-            "18"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Age Over 18"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "age_in_years"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Age in Years"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "age_birth_year"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Age Year of Birth"
-            }
-          ],
-          "sd": "always"
-        },
-        {
-          "path": [
-            "trust_anchor"
-          ],
-          "display": [
-            {
-              "lang": "en",
-              "label": "Trust Anchor"
-            }
-          ],
-          "sd": "always"
-        }
-      ]
-      
-    }
-""".trimIndent()
